embed_documents/layer_file_utils/requirements.txt
boto3>=1.26.0
python-docx>=0.8.11
PyPDF2>=3.0.0
typing-extensions>=4.5.0
pandas>=2.0.0
openpyxl>=3.1.0
python-pptx>=0.6.21
langchain>=0.1.0
requests>=2.31.0
